Final_Version/Patching_and_Blending/requirements.txt
absl-py==2.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
exceptiongroup==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.110.0
fastdeploy-python==1.0.7
fastdeploy-tools==0.0.5
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2023.4.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
jax==0.4.25
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.42.0
Mako==1.3.2
MarkupSafe==2.1.3
matplotlib==3.8.3
mediapipe==0.10.10
ml-dtypes==0.3.2
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optuna==3.5.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==3.20.3
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
SQLAlchemy==2.0.28
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
umap==0.1.1
umap-learn==0.5.5
urllib3==2.2.1
uvicorn==0.16.0
